+++ sphinx/source/changelog.rst
@@ -3,8 +3,9 @@ Changelog
 
 6.2b1
 -----
-* The ``--org``, ``--stack`` and ``--start`` options of :ref:`bin2sna.py` now
-  accept a hexadecimal integer prefixed by '0x'
+* The ``--clear``, ``--end``, ``--org``, ``--stack`` and ``--start`` options of
+  :ref:`bin2sna.py` and :ref:`bin2tap.py` now accept a hexadecimal integer
+  prefixed by '0x'
 * Added the ``--show-config`` option to :ref:`skool2asm.py`,
   :ref:`skool2html.py` and :ref:`sna2skool.py` (for showing configuration
   parameter values)